+++ src/Controller/DaftarController.java
@@ -3,8 +3,13 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package sistem_lowongan_kerja;
+package Controller;
 
+import Model.Model;
+import View.DaftarPelamarpage;
+import View.DaftarPerusahaanpage;
+import View.Daftarpage;
+import View.Homepage;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 import javax.swing.JOptionPane;
