+++ src/Controller/LoginController.java
@@ -3,8 +3,12 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package sistem_lowongan_kerja;
+package Controller;
 
+import Model.Model;
+import View.Daftarpage;
+import View.Homepage;
+import View.Loginpage;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 import javax.swing.JOptionPane;
