+++ src/Controller/PelamarSetController.java
@@ -1,0 +1,23 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package Controller;
+
+import Model.Model;
+import View.PelamarSetpage;
+
+/**
+ *
+ * @author Diah Ajeng
+ */
+public class PelamarSetController {
+    private PelamarSetpage view;
+    private Model model;
+
+    public PelamarSetController(PelamarSetpage view, Model model) {
+        this.view = view;
+        this.model = model;
+    }
+}
