+++ src/Data/Lowongan.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package sistem_lowongan_kerja;
+package Data;
 
 import java.util.Date;
 import java.util.List;
