+++ src/View/DaftarPelamarpage.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package sistem_lowongan_kerja;
+package View;
 
 import java.awt.event.ActionListener;
 
@@ -228,15 +228,15 @@ public class DaftarPelamarpage extends javax.swing.JFrame {
         });
     }
 
-    void addListenbatal(ActionListener listenbtnbatal){
+    public void addListenbatal(ActionListener listenbtnbatal){
         btnbatal.addActionListener(listenbtnbatal);
     }
     
-    void addListendaftar(ActionListener listenbtndaftar){
+    public void addListendaftar(ActionListener listenbtndaftar){
         btndaftar.addActionListener(listenbtndaftar);
     }
     
-    void addListenkeluar(ActionListener listenbtnkeluar){
+    public void addListenkeluar(ActionListener listenbtnkeluar){
         btnkeluar.addActionListener(listenbtnkeluar);
     }
