--- src/sistem_lowongan_kerja/CariController.java
@@ -1,22 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package sistem_lowongan_kerja.Controller;
-
-import sistem_lowongan_kerja.Caripage;
-import sistem_lowongan_kerja.Model;
-
-/**
- *
- * @author Diah Ajeng
- */
-public class CariController {
-    private Caripage view;
-    private Model model;
-    
-    public CariController(Caripage view, Model model){
-        
-    }
-}
